Digital Image Processing, 4th EDITION, Gonzalez
 This edition of Digital Image Processing is a major revision of the book. As in 
the 1977 and 1987 editions by Gonzalez and Wintz, and the 1992, 2002, and 2008  
editions by Gonzalez and Woods, this sixth-generation edition was prepared 
with students and instructors in mind. The principal objectives of the book 
continue to be to provide an introduction to basic concepts and methodologies 
applicable to digital image processing, and to develop a foundation that can 
be used as the basis for further study and research in this field.